third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/tab-size/tab-size-expected.txt

This is a testharness.js-based test.
FAIL Test tab-size measurements assert_approx_equals: expected 4.03125 +/- 0.02 but got 23.40625
Harness: the test ran to completion.
